更新数据库列类型，优化/f消息格式

+++ src/main/java/xyz/zcraft/command/resolution/UidResolution.java
@@ -1,4 +1,4 @@
 package xyz.zcraft.command.resolution;
 
-public record UidResolution(Integer uid, String errorMessage) {
+public record UidResolution(Long uid, String errorMessage) {
 }
